+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix crash on IntelliJ 2025.1 caused by `IconUtil.colorize` signature change in 2025.2
+
 ## [0.3.0] - 2026-02-24
 
 ### Added